+++ CHANGELOG.md
@@ -41,6 +41,8 @@ All notable changes to eww will be listed here, starting at changes since versio
 - Add `:lines` and `:wrap-mode` properties to label widget (By: vaporii)
 - Add `value-pos` to scale widget (By: ipsvn)
 - Add `floor` and `ceil` function calls to simplexpr (By: wsbankenstein)
+- Add `uppercase` and `lowercase` functions for expressions (By: dod-101)
+
 
 ## [0.6.0] (21.04.2024)